moved functions into _helpers script
